Changes in some classes and Pom.xml

+++ CourseManagementBackend/src/main/java/com/app/controller/SubjectController.java
@@ -3,7 +3,6 @@ package com.app.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -49,20 +48,20 @@ public class SubjectController {
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<Subject> getSubjectById(@PathVariable Long id)
+	public ResponseEntity<Subject> getSubjectById(@PathVariable int id)
 	{
 		return ResponseEntity.ok(subjectService.getSubjectById(id));
 	}
 	
 	
 	@PutMapping("/{id}")
-	public Subject updateSubject(@PathVariable Long id, @RequestBody SubjectRespDto dto) {		
+	public Subject updateSubject(@PathVariable int id, @RequestBody SubjectRespDto dto) {		
 		return subjectService.updateSubject(id, dto);
 	}
 	
 	
 	@DeleteMapping("/{id}")
-	public void deleteSubject(@PathVariable Long id)
+	public void deleteSubject(@PathVariable int id)
 		{
 			subjectService.deleteSubject(id);
 		}
